Updated login with backend with static values
